refactor: rename getLocation to getPointer
